+++ node-src/types.ts
@@ -359,6 +359,7 @@ export interface Task {
 }
 
 export interface Reason {
+  resolvedModule?: string; // rspack sets a resolvedModule field that holds the module name
   moduleName: string;
 }
 export interface Module {